In order to achieve more plugin support
