+++ dolphinscheduler-task-plugin/dolphinscheduler-task-datax/src/main/java/org/apache/dolphinscheduler/plugin/task/datax/DataxTask.java
@@ -17,14 +17,10 @@
 
 package org.apache.dolphinscheduler.plugin.task.datax;
 
-import static org.apache.dolphinscheduler.plugin.datasource.api.utils.PasswordUtils.decodePassword;
 import static org.apache.dolphinscheduler.plugin.task.api.TaskConstants.EXIT_CODE_FAILURE;
-import static org.apache.dolphinscheduler.spi.enums.DbType.DORIS;
 
 import org.apache.dolphinscheduler.common.log.SensitiveDataConverter;
 import org.apache.dolphinscheduler.common.utils.JSONUtils;
-import org.apache.dolphinscheduler.plugin.datasource.api.plugin.DataSourceClientProvider;
-import org.apache.dolphinscheduler.plugin.datasource.api.utils.DataSourceUtils;
 import org.apache.dolphinscheduler.plugin.task.api.AbstractTask;
 import org.apache.dolphinscheduler.plugin.task.api.ShellCommandExecutor;
 import org.apache.dolphinscheduler.plugin.task.api.TaskCallBack;
@@ -36,37 +32,23 @@ import org.apache.dolphinscheduler.plugin.task.api.parameters.AbstractParameters
 import org.apache.dolphinscheduler.plugin.task.api.shell.IShellInterceptorBuilder;
 import org.apache.dolphinscheduler.plugin.task.api.shell.ShellInterceptorBuilderFactory;
 import org.apache.dolphinscheduler.plugin.task.api.utils.ParameterUtils;
+import org.apache.dolphinscheduler.plugin.task.datax.content.BuildDataxJobContentJsonFactory;
+import org.apache.dolphinscheduler.plugin.task.datax.content.reader.AbstractBuildDataxJobContentJsonReader;
+import org.apache.dolphinscheduler.plugin.task.datax.content.reader.BuildDataxJobContentJsonDefaultReader;
+import org.apache.dolphinscheduler.plugin.task.datax.content.writer.AbstractBuildDataxJobContentJsonWriter;
 import org.apache.dolphinscheduler.spi.datasource.BaseConnectionParam;
-import org.apache.dolphinscheduler.spi.enums.DbType;
 import org.apache.dolphinscheduler.spi.enums.Flag;
 
-import org.apache.commons.collections4.CollectionUtils;
 import org.apache.commons.io.FileUtils;
 
 import java.io.File;
 import java.nio.charset.StandardCharsets;
 import java.nio.file.Files;
 import java.nio.file.Path;
-import java.sql.Connection;
-import java.sql.PreparedStatement;
-import java.sql.ResultSet;
-import java.sql.ResultSetMetaData;
-import java.sql.SQLException;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
-import java.util.concurrent.ExecutionException;
 
-import com.alibaba.druid.sql.ast.SQLStatement;
-import com.alibaba.druid.sql.ast.expr.SQLIdentifierExpr;
-import com.alibaba.druid.sql.ast.expr.SQLPropertyExpr;
-import com.alibaba.druid.sql.ast.statement.SQLSelect;
-import com.alibaba.druid.sql.ast.statement.SQLSelectItem;
-import com.alibaba.druid.sql.ast.statement.SQLSelectQueryBlock;
-import com.alibaba.druid.sql.ast.statement.SQLSelectStatement;
-import com.alibaba.druid.sql.ast.statement.SQLUnionQuery;
-import com.alibaba.druid.sql.parser.SQLStatementParser;
-import com.fasterxml.jackson.databind.node.ArrayNode;
 import com.fasterxml.jackson.databind.node.ObjectNode;
 
 public class DataxTask extends AbstractTask {
@@ -82,10 +64,7 @@ public class DataxTask extends AbstractTask {
      */
     private static final String PYTHON_LAUNCHER = "${PYTHON_LAUNCHER}";
 
-    /**
-     * select all
-     */
-    private static final String SELECT_ALL_CHARACTER = "*";
+    BuildDataxJobContentJsonFactory buildDataxJobContentJsonFactory;
 
     /**
      * post jdbc info regex
@@ -235,109 +214,21 @@ public class DataxTask extends AbstractTask {
      * @return collection of datax job config JSONObject
      */
     private List<ObjectNode> buildDataxJobContentJson() {
-        BaseConnectionParam dataSourceCfg = (BaseConnectionParam) DataSourceUtils.buildConnectionParams(
-                dataxTaskExecutionContext.getSourcetype(),
-                dataxTaskExecutionContext.getSourceConnectionParams());
+        buildDataxJobContentJsonFactory = new BuildDataxJobContentJsonFactory();
+        AbstractBuildDataxJobContentJsonReader reader = buildDataxJobContentJsonFactory.getReader(dataxTaskExecutionContext.getSourcetype());
+        AbstractBuildDataxJobContentJsonWriter writer = buildDataxJobContentJsonFactory.getWriter(dataxTaskExecutionContext.getTargetType());
 
-        BaseConnectionParam dataTargetCfg = (BaseConnectionParam) DataSourceUtils.buildConnectionParams(
-                dataxTaskExecutionContext.getTargetType(),
-                dataxTaskExecutionContext.getTargetConnectionParams());
+        reader.init(dataxTaskExecutionContext,dataXParameters);
+        writer.init(dataxTaskExecutionContext,dataXParameters);
 
-        ObjectNode reader;
-        ObjectNode writer;
-
-        switch (dataxTaskExecutionContext.getSourcetype()) {
-            default:
-                reader = buildDataxJobContentJsonDefaultReader(dataSourceCfg);
-        }
-
-        switch (dataxTaskExecutionContext.getTargetType()) {
-            case DORIS:
-            default:
-                writer = buildDataxJobContentJsonDefaultWrite(dataSourceCfg, dataTargetCfg);
-        }
 
         List<ObjectNode> contentList = new ArrayList<>();
         ObjectNode content = JSONUtils.createObjectNode();
-        content.set("reader", reader);
-        content.set("writer", writer);
+        content.set("reader", reader.reader());
+        content.set("writer", writer.writer());
         contentList.add(content);
 
         return contentList;
-    }
-
-    private ObjectNode buildDataxJobContentJsonDefaultReader(BaseConnectionParam dataSourceCfg) {
-
-        List<ObjectNode> readerConnArr = new ArrayList<>();
-        ObjectNode readerConn = JSONUtils.createObjectNode();
-
-        ArrayNode sqlArr = readerConn.putArray("querySql");
-        for (String sql : new String[]{dataXParameters.getSql()}) {
-            sqlArr.add(sql);
-        }
-
-        ArrayNode urlArr = readerConn.putArray("jdbcUrl");
-        urlArr.add(DataSourceUtils.getJdbcUrl(DbType.valueOf(dataXParameters.getDsType()), dataSourceCfg));
-
-        readerConnArr.add(readerConn);
-
-        ObjectNode readerParam = JSONUtils.createObjectNode();
-        readerParam.put("username", dataSourceCfg.getUser());
-        readerParam.put("password", decodePassword(dataSourceCfg.getPassword()));
-        readerParam.putArray("connection").addAll(readerConnArr);
-
-        ObjectNode reader = JSONUtils.createObjectNode();
-        reader.put("name", DataxUtils.getReaderPluginName(dataxTaskExecutionContext.getSourcetype()));
-        reader.set("parameter", readerParam);
-
-        return reader;
-    }
-
-    private ObjectNode buildDataxJobContentJsonDefaultWrite(BaseConnectionParam dataSourceCfg,
-                                                            BaseConnectionParam dataTargetCfg) {
-
-        List<ObjectNode> writerConnArr = new ArrayList<>();
-        ObjectNode writerConn = JSONUtils.createObjectNode();
-        ArrayNode tableArr = writerConn.putArray("table");
-        tableArr.add(dataXParameters.getTargetTable());
-
-        writerConn.put("jdbcUrl",
-                DataSourceUtils.getJdbcUrl(DbType.valueOf(dataXParameters.getDtType()), dataTargetCfg));
-        writerConnArr.add(writerConn);
-
-        ObjectNode writerParam = JSONUtils.createObjectNode();
-        writerParam.put("username", dataTargetCfg.getUser());
-        writerParam.put("password", decodePassword(dataTargetCfg.getPassword()));
-
-        String[] columns = parsingSqlColumnNames(dataxTaskExecutionContext.getSourcetype(),
-                dataxTaskExecutionContext.getTargetType(),
-                dataSourceCfg, dataXParameters.getSql());
-
-        ArrayNode columnArr = writerParam.putArray("column");
-        for (String column : columns) {
-            columnArr.add(column);
-        }
-        writerParam.putArray("connection").addAll(writerConnArr);
-
-        if (CollectionUtils.isNotEmpty(dataXParameters.getPreStatements())) {
-            ArrayNode preSqlArr = writerParam.putArray("preSql");
-            for (String preSql : dataXParameters.getPreStatements()) {
-                preSqlArr.add(preSql);
-            }
-
-        }
-
-        if (CollectionUtils.isNotEmpty(dataXParameters.getPostStatements())) {
-            ArrayNode postSqlArr = writerParam.putArray("postSql");
-            for (String postSql : dataXParameters.getPostStatements()) {
-                postSqlArr.add(postSql);
-            }
-        }
-
-        ObjectNode writer = JSONUtils.createObjectNode();
-        writer.put("name", DataxUtils.getWriterPluginName(dataxTaskExecutionContext.getTargetType()));
-        writer.set("parameter", writerParam);
-        return writer;
     }
 
     /**
@@ -431,146 +322,9 @@ public class DataxTask extends AbstractTask {
         return String.format(JVM_PARAM, xms, xmx);
     }
 
-    /**
-     * parsing synchronized column names in SQL statements
-     *
-     * @param sourceType the database type of the data source
-     * @param targetType the database type of the data target
-     * @param dataSourceCfg the database connection parameters of the data source
-     * @param sql sql for data synchronization
-     * @return Keyword converted column names
-     */
-    private String[] parsingSqlColumnNames(DbType sourceType, DbType targetType, BaseConnectionParam dataSourceCfg,
-                                           String sql) {
-        String[] columnNames = tryGrammaticalAnalysisSqlColumnNames(sourceType, sql);
-
-        if (columnNames == null || columnNames.length == 0) {
-            log.info("try to execute sql analysis query column name");
-            columnNames = tryExecuteSqlResolveColumnNames(sourceType, dataSourceCfg, sql);
-        }
-
-        notNull(columnNames, String.format("parsing sql columns failed : %s", sql));
-
-        return DataxUtils.convertKeywordsColumns(targetType, columnNames);
-    }
-
-    /**
-     * try grammatical parsing column
-     *
-     * @param dbType database type
-     * @param sql sql for data synchronization
-     * @return column name array
-     * @throws RuntimeException if error throws RuntimeException
-     */
-    private String[] tryGrammaticalAnalysisSqlColumnNames(DbType dbType, String sql) {
-        String[] columnNames;
-
-        try {
-            SQLStatementParser parser = DataxUtils.getSqlStatementParser(dbType, sql);
-            if (parser == null) {
-                log.warn("database driver [{}] is not support grammatical analysis sql", dbType);
-                return new String[0];
-            }
-
-            SQLStatement sqlStatement = parser.parseStatement();
-            SQLSelectStatement sqlSelectStatement = (SQLSelectStatement) sqlStatement;
-            SQLSelect sqlSelect = sqlSelectStatement.getSelect();
-
-            List<SQLSelectItem> selectItemList = null;
-            if (sqlSelect.getQuery() instanceof SQLSelectQueryBlock) {
-                SQLSelectQueryBlock block = (SQLSelectQueryBlock) sqlSelect.getQuery();
-                selectItemList = block.getSelectList();
-            } else if (sqlSelect.getQuery() instanceof SQLUnionQuery) {
-                SQLUnionQuery unionQuery = (SQLUnionQuery) sqlSelect.getQuery();
-                SQLSelectQueryBlock block = (SQLSelectQueryBlock) unionQuery.getRight();
-                selectItemList = block.getSelectList();
-            }
-
-            notNull(selectItemList,
-                    String.format("select query type [%s] is not support", sqlSelect.getQuery().toString()));
-
-            columnNames = new String[selectItemList.size()];
-            for (int i = 0; i < selectItemList.size(); i++) {
-                SQLSelectItem item = selectItemList.get(i);
-
-                String columnName = null;
-
-                if (item.getAlias() != null) {
-                    columnName = item.getAlias();
-                } else if (item.getExpr() != null) {
-                    if (item.getExpr() instanceof SQLPropertyExpr) {
-                        SQLPropertyExpr expr = (SQLPropertyExpr) item.getExpr();
-                        columnName = expr.getName();
-                    } else if (item.getExpr() instanceof SQLIdentifierExpr) {
-                        SQLIdentifierExpr expr = (SQLIdentifierExpr) item.getExpr();
-                        columnName = expr.getName();
-                    }
-                } else {
-                    throw new RuntimeException(
-                            String.format("grammatical analysis sql column [ %s ] failed", item));
-                }
-
-                if (SELECT_ALL_CHARACTER.equals(item.toString())) {
-                    log.info("sql contains *, grammatical analysis failed");
-                    return new String[0];
-                }
-
-                if (columnName == null) {
-                    throw new RuntimeException(
-                            String.format("grammatical analysis sql column [ %s ] failed", item));
-                }
-
-                columnNames[i] = columnName;
-            }
-        } catch (Exception e) {
-            log.warn(e.getMessage(), e);
-            return new String[0];
-        }
-
-        return columnNames;
-    }
-
-    /**
-     * try to execute sql to resolve column names
-     *
-     * @param baseDataSource the database connection parameters
-     * @param sql sql for data synchronization
-     * @return column name array
-     */
-    public String[] tryExecuteSqlResolveColumnNames(DbType sourceType, BaseConnectionParam baseDataSource, String sql) {
-        String[] columnNames;
-        sql = String.format("SELECT t.* FROM ( %s ) t WHERE 0 = 1", sql);
-        sql = sql.replace(";", "");
-
-        try (
-                Connection connection =
-                        DataSourceClientProvider.getAdHocConnection(sourceType, baseDataSource);
-                PreparedStatement stmt = connection.prepareStatement(sql);
-                ResultSet resultSet = stmt.executeQuery()) {
-
-            ResultSetMetaData md = resultSet.getMetaData();
-            int num = md.getColumnCount();
-            columnNames = new String[num];
-            for (int i = 1; i <= num; i++) {
-                columnNames[i - 1] = md.getColumnName(i).replace("t.", "");
-            }
-        } catch (SQLException | ExecutionException e) {
-            log.error(e.getMessage(), e);
-            return null;
-        }
-
-        return columnNames;
-    }
-
     @Override
     public AbstractParameters getParameters() {
         return dataXParameters;
-    }
-
-    private void notNull(Object obj, String message) {
-        if (obj == null) {
-            throw new RuntimeException(message);
-        }
     }
 
 }
