+++ dolphinscheduler-task-plugin/dolphinscheduler-task-datax/src/main/java/org/apache/dolphinscheduler/plugin/task/datax/content/BuildDataxJobContentJsonFactory.java
@@ -1,0 +1,24 @@
+package org.apache.dolphinscheduler.plugin.task.datax.content;
+
+
+import org.apache.dolphinscheduler.plugin.task.datax.content.reader.AbstractBuildDataxJobContentJsonReader;
+import org.apache.dolphinscheduler.plugin.task.datax.content.reader.BuildDataxJobContentJsonDefaultReader;
+import org.apache.dolphinscheduler.plugin.task.datax.content.writer.AbstractBuildDataxJobContentJsonWriter;
+import org.apache.dolphinscheduler.plugin.task.datax.content.writer.BuildDataxJobContentJsonDefaultWriter;
+import org.apache.dolphinscheduler.spi.enums.DbType;
+
+public class BuildDataxJobContentJsonFactory {
+    public AbstractBuildDataxJobContentJsonReader getReader(DbType dbType){
+        switch (dbType){
+            default: new BuildDataxJobContentJsonDefaultReader();
+        }
+        return null;
+    }
+
+    public AbstractBuildDataxJobContentJsonWriter getWriter(DbType dbType){
+        switch (dbType){
+            default: new BuildDataxJobContentJsonDefaultWriter();
+        }
+        return null;
+    }
+}
